+++ crates/runtime/benches/bench_spicecloud/tpch_q12.sql
