+++ crates/runtime/benches/bench_spicecloud/tpch_q16.sql
