+++ crates/runtime/benches/bench_spicecloud/tpch_q17.sql
