+++ crates/runtime/benches/bench_spicecloud/tpch_q18.sql
