+++ crates/runtime/benches/bench_spicecloud/tpch_q2.sql
