+++ crates/runtime/benches/bench_spicecloud/tpch_q20.sql
