+++ crates/runtime/benches/bench_spicecloud/tpch_q21.sql
