+++ crates/runtime/benches/bench_spicecloud/tpch_q22.sql
