+++ crates/runtime/benches/bench_spicecloud/tpch_q3.sql
