+++ crates/runtime/benches/bench_spicecloud/tpch_q6.sql
