+++ crates/runtime/benches/bench_spicecloud/tpch_q8.sql
