+++ crates/runtime/benches/bench_spicecloud/tpch_q9.sql
