+++ crates/runtime/benches/bench_spicecloud/tpch_simple_q1.sql
