+++ crates/runtime/benches/bench_spicecloud/tpch_simple_q3.sql
