+++ crates/runtime/benches/bench_spicecloud/tpch_simple_q5.sql
